pyproject.toml
[project]
name = "repl-nix-glaucomatriageai"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.10",
    "pillow>=11.0.0",
    "numpy>=2.1.3",
    "torch>=2.5.1",
    "transformers>=4.46.3",
    "accelerate>=1.1.1",
]
